third_party/blink/web_tests/external/wpt/css/css-pseudo/first-letter-allowed-properties-expected.txt

This is a testharness.js-based test.
[FAIL] textJustify should be applied to first-letter pseudo elements.
  assert_equals: expected "inter-word" but got "auto"
Harness: the test ran to completion.
